spike-platform/apps/api/pyproject.toml
[project]
name = "spike-api"
version = "0.1.0"
description = "SPIKE Platform - AI Wealth Intelligence API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "redis>=5.0.1",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.26.0",

    # Data Processing
    "pandas>=2.1.4",
    "numpy>=1.26.3",
    "yfinance>=0.2.35",

    # AI/ML
    "openai>=1.8.0",
    "anthropic>=0.10.0",
    "torch>=2.1.2",

    # Utilities
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.10",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
